fix: split oversized mortgage and project-detail chunks under 20KB budget
Extract mortgage success CTA + doc upload into mortgage-success.ts (lazy-loaded
on step 4 submission) and project lightbox into project-lightbox.ts (lazy-loaded
on first image tap). Brings mortgage chunk from 20.6KB→16.5KB and project-detail
from 22.5KB→19.2KB.
